refactor: реорганизация хранилищ

+++ src/main/java/ru/yandex/practicum/filmorate/service/film/FilmService.java
@@ -9,6 +9,8 @@ import ru.yandex.practicum.filmorate.exceptions.NotFoundException;
 import ru.yandex.practicum.filmorate.exceptions.ValidationException;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.storage.film.FilmStorage;
+import ru.yandex.practicum.filmorate.storage.mpa.MpaStorage;
+import ru.yandex.practicum.filmorate.storage.user.UserStorage;
 
 import java.time.LocalDate;
 import java.util.List;
@@ -20,6 +22,14 @@ public class FilmService {
     @Autowired
     @Qualifier("dbFilmStorage")
     private FilmStorage filmStorage;
+
+    @Autowired
+    @Qualifier("dbUserStorage")
+    private UserStorage userStorage;
+
+    @Autowired
+    @Qualifier("dbMpaStorage")
+    private MpaStorage mpaStorage;
 
     /*
     Список операций:
@@ -41,7 +51,13 @@ public class FilmService {
     public Film updateFilm(Film film) throws ValidationException, NotFoundException {
         log.debug("Запрос на изменение фильма: {}", film);
         validate(film);
-        return filmStorage.updateFilm(film);
+
+        Film dbFilm = filmStorage.updateFilm(film);
+        if (dbFilm == null) {
+            log.info("Фильм с указанным ID {} не найден в базе данных", film.getId());
+            throw new NotFoundException(String.format("Фильм с указанным ID (%d) не найден", film.getId()));
+        }
+        return dbFilm;
     }
 
     public Film getFilm(Long id) throws NotFoundException {
@@ -59,11 +75,27 @@ public class FilmService {
 
     public void addUserLike(Long filmId, Long userId) throws NotFoundException, ValidationException {
         log.debug("Запрос на установку лайка фильму: filmId={}, userId={}", filmId, userId);
+        if (filmStorage.getFilm(filmId) == null) {
+            log.info("Фильм с указанным ID {} не найден в базе данных", filmId);
+            throw new ValidationException(String.format("Фильм с указанным ID %d не найден в базе данных", filmId));
+        }
+        if (userStorage.getUser(userId) == null) {
+            log.info("Пользователь с указанным ID {} не найден в базе данных", userId);
+            throw new ValidationException(String.format("Пользователь с указанным ID %d не найден в базе данных", userId));
+        }
         filmStorage.addUserLike(filmId, userId);
     }
 
     public void deleteUserLike(Long filmId, Long userId) throws NotFoundException, ValidationException {
         log.debug("Запрос на снятие лайка фильму: filmId={}, userId={}", filmId, userId);
+        if (filmStorage.getFilm(filmId) == null) {
+            log.info("Фильм с указанным ID {} не найден в базе данных", filmId);
+            throw new ValidationException(String.format("Фильм с указанным ID %d не найден в базе данных", filmId));
+        }
+        if (userStorage.getUser(userId) == null) {
+            log.info("Пользователь с указанным ID {} не найден в базе данных", userId);
+            throw new ValidationException(String.format("Пользователь с указанным ID %d не найден в базе данных", userId));
+        }
         filmStorage.deleteUserLike(filmId, userId);
     }
 
@@ -93,6 +125,13 @@ public class FilmService {
         if (film.getDuration() == null || film.getDuration() <= 0) {
             throw new ValidationException("Некорректная продолжительность фильма");
         }
+        if (film.getMpa() != null && film.getMpa().getId() != null) {
+            int mpaId = film.getMpa().getId();
+            if (mpaStorage.getMpa(mpaId) == null) {
+                throw new ValidationException(String.format("MPA-рейтинг с указанным ID %d не найден в базе данных", mpaId));
+            }
+        }
+
     }
 
 }
